backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.5
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0 